refactor: refact table examples
