--- src/app/routes/tables/basic/basic.component.html
@@ -1,9 +1,0 @@
-<page-header class="bg-green-600"></page-header>
-
-<mtx-grid [data]="list"
-          [columns]="columns"
-          [loading]="isLoading"
-          [pageSizeOptions]="[5,10,50,100]"
-          (sortChange)="changeSort($event)"
-          (selectionChange)="changeSelect($event)">
-</mtx-grid>
