+++ src/app/routes/tables/ktchen-sink/ktchen-sink.service.ts
@@ -1,0 +1,6 @@
+import { Injectable } from '@angular/core';
+
+@Injectable()
+export class TablesKtchenSinkService {
+  constructor() {}
+}
