+++ src/app/routes/tables/remote-data/remote-data.component.ts
@@ -1,0 +1,92 @@
+import { Component, OnInit, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
+
+import { MtxGridColumn } from '@ng-matero/extensions';
+import { TablesRemoteDataService } from './remote-data.service';
+import { PageEvent } from '@angular/material/paginator';
+import { delEmptyKey } from '@shared';
+
+@Component({
+  selector: 'app-tables-remote-data',
+  templateUrl: './remote-data.component.html',
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  providers: [TablesRemoteDataService],
+})
+export class TablesRemoteDataComponent implements OnInit {
+  columns: MtxGridColumn[] = [
+    {
+      header: 'Name',
+      field: 'name',
+      formatter: (data: any) => `<a href="${data.html_url}" target="_blank">${data.name}</a>`,
+    },
+    { header: 'Owner', field: 'owner.login' },
+    { header: 'Owner Avatar', field: 'owner.avatar_url', type: 'image' },
+    { header: 'Description', field: 'description', width: '300px' },
+    { header: 'stars', field: 'stargazers_count' },
+    { header: 'forks', field: 'forks_count' },
+    { header: 'Score', field: 'score' },
+    { header: 'Issues', field: 'open_issues' },
+    { header: 'Language', field: 'language' },
+    { header: 'License', field: 'license.name' },
+    { header: 'Home Page', field: 'homepage', type: 'link' },
+    {
+      header: 'Is forked',
+      field: 'fork',
+      formatter: (data: any) => JSON.stringify(data.fork),
+    },
+    {
+      header: 'Archived',
+      field: 'archived',
+      type: 'tag',
+      tag: {
+        true: { text: 'Yes', color: 'red-100' },
+        false: { text: 'No', color: 'green-100' },
+      },
+    },
+    { header: 'Created Date', field: 'created_at' },
+    { header: 'Updated Date', field: 'updated_at' },
+  ];
+  list = [];
+  total = 0;
+  isLoading = true;
+
+  query = {
+    q: 'user:nzbin',
+    sort: 'stars',
+    order: 'desc',
+    page: 0,
+    per_page: 5,
+  };
+
+  get params() {
+    const p = Object.assign({}, this.query);
+    p.page += 1;
+    return delEmptyKey(p);
+  }
+
+  constructor(private remoteSrv: TablesRemoteDataService, private cdr: ChangeDetectorRef) {}
+
+  ngOnInit() {
+    this.getData();
+  }
+
+  getData() {
+    this.isLoading = true;
+    this.remoteSrv.getData(this.params).subscribe((res: any) => {
+      this.list = res.items;
+      this.total = res.total_count;
+      this.isLoading = false;
+      this.cdr.detectChanges();
+    });
+  }
+
+  getNextPage(e: PageEvent) {
+    this.query.page = e.pageIndex;
+    this.query.per_page = e.pageSize;
+    this.getData();
+  }
+
+  search() {
+    this.query.page = 0;
+    this.getData();
+  }
+}
