+++ src/app/routes/tables/tables.module.ts
@@ -2,12 +2,12 @@ import { NgModule } from '@angular/core';
 import { SharedModule } from '@shared/shared.module';
 import { TableRoutingModule } from './tables-routing.module';
 
-import { TableBasicComponent } from './basic/basic.component';
-import { TableBasicEditComponent } from './basic/edit/edit.component';
-import { TablesAdvancedComponent } from './advanced/advanced.component';
+import { TablesKtchenSinkComponent } from './ktchen-sink/ktchen-sink.component';
+import { TablesKtchenSinkEditComponent } from './ktchen-sink/edit/edit.component';
+import { TablesRemoteDataComponent } from './remote-data/remote-data.component';
 
-const COMPONENTS = [TableBasicComponent, TablesAdvancedComponent];
-const COMPONENTS_DYNAMIC = [TableBasicEditComponent];
+const COMPONENTS = [TablesKtchenSinkComponent, TablesRemoteDataComponent];
+const COMPONENTS_DYNAMIC = [TablesKtchenSinkEditComponent];
 
 @NgModule({
   imports: [SharedModule, TableRoutingModule],
